+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 
 - [Feature: Add config change listener feature support](https://github.com/Tencent/spring-cloud-tencent/pull/254)
+- [Feature: Add config module unit test](https://github.com/Tencent/spring-cloud-tencent/pull/255)